requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
shap==0.41.0
lime==0.2.0.1
faker==18.10.1
streamlit==1.22.0
fastapi==0.95.1
uvicorn==0.21.1
python-multipart==0.0.6
plotly==5.14.1
nltk==3.8.1
joblib==1.2.0
pyyaml==6.0